# wiki/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ops Agent API",
    "version": "0.1.0",
    "description": "Generated from registered Ops Agent routes. Middleware authentication, manually parsed request models and SSE media types are documented. WebSocket frames are described separately. Documentation viewer does not execute requests."
  },
  "paths": {
...
    "/api/model-configs": {
      "get": {
        "summary": "List Model Config Records",
        "operationId": "list_model_config_records_api_model_configs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelConfigView"
                  },
                  "type": "array",
                  "title": "Response List Model Config Records Api Model Configs Get"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      },
      "post": {
        "summary": "Create Model Config Record",
        "operationId": "create_model_config_record_api_model_configs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      }
    },
    "/api/model-configs/{config_id}": {
      "get": {
        "summary": "Get Model Config Record",
        "operationId": "get_model_config_record_api_model_configs__config_id__get",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      },
      "put": {
        "summary": "Update Model Config Record",
        "operationId": "update_model_config_record_api_model_configs__config_id__put",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      },
      "delete": {
        "summary": "Delete Model Config Record",
        "operationId": "delete_model_config_record_api_model_configs__config_id__delete",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      }
    },
    "/api/model-configs/{config_id}/default": {
      "post": {
        "summary": "Set Default Model Config Record",
        "operationId": "set_default_model_config_record_api_model_configs__config_id__default_post",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      }
    },
    "/api/model-configs/discover": {
      "post": {
        "summary": "Discover Model Configs",
        "operationId": "discover_model_configs_api_model_configs_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDiscoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDiscoveryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      }
    },
    "/api/model-configs/test": {
      "post": {
        "summary": "Test Model Config",
        "operationId": "test_model_config_api_model_configs_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConnectionTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConnectionTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "model-configs"
        ]
      }
    },
    "/api/network-topology/snapshots": {
      "get": {
        "tags": [
          "network-topology"
        ],
        "summary": "List Snapshots",
        "operationId": "list_snapshots_api_network_topology_snapshots_get",
        "parameters": [
          {
            "name": "instance_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instance Id"
            }
          },
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Organization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Snapshots Api Network Topology Snapshots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "network-topology"
        ],
        "summary": "Collect Snapshot",
        "operationId": "collect_snapshot_api_network_topology_snapshots_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopologyCollectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Collect Snapshot Api Network Topology Snapshots Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/network-topology/snapshots/{snapshot_id}": {
      "get": {
        "tags": [
          "network-topology"
        ],
        "summary": "Get Snapshot",
        "operationId": "get_snapshot_api_network_topology_snapshots__snapshot_id__get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Snapshot Api Network Topology Snapshots  Snapshot Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelConfigCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "API 服务"
          },
          "provider": {
            "type": "string",
            "const": "anthropic",
            "title": "Provider",
            "default": "anthropic"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "api_key": {
            "type": "string",
            "format": "password",
            "title": "Api Key",
            "writeOnly": true
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": ""
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          },
          "timeout_seconds": {
            "type": "integer",
            "title": "Timeout Seconds",
            "default": 30
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.2
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "default": 1024
          },
          "prompt_cache_enabled": {
            "type": "boolean",
            "title": "Prompt Cache Enabled",
            "default": true
          },
          "prompt_cache_ttl": {
            "type": "string",
            "enum": [
              "ephemeral",
              "one_hour"
            ],
            "title": "Prompt Cache Ttl",
            "default": "ephemeral"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "base_url",
          "api_key"
        ],
        "title": "ModelConfigCreate"
      },
      "ModelConfigUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "const": "anthropic"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "prompt_cache_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Cache Enabled"
          },
          "prompt_cache_ttl": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ephemeral",
                  "one_hour"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Cache Ttl"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "ModelConfigUpdate"
      },
      "ModelConfigView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "api_key_masked": {
            "type": "string",
            "title": "Api Key Masked"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "timeout_seconds": {
            "type": "integer",
            "title": "Timeout Seconds",
            "default": 30
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.2
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "default": 1024
          },
          "prompt_cache_enabled": {
            "type": "boolean",
            "title": "Prompt Cache Enabled",
            "default": true
          },
          "prompt_cache_ttl": {
            "type": "string",
            "enum": [
              "ephemeral",
              "one_hour"
            ],
            "title": "Prompt Cache Ttl",
            "default": "ephemeral"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "provider",
          "base_url",
          "api_key_masked",
          "model_name",
          "is_default",
          "description"
        ],
        "title": "ModelConfigView"
      },
      "ModelConnectionTestRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "const": "anthropic",
            "title": "Provider"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "api_key": {
            "type": "string",
            "format": "password",
            "title": "Api Key",
            "writeOnly": true
          },
          "model_name": {
            "type": "string",
            "minLength": 1,
            "title": "Model Name"
          },
          "timeout_seconds": {
            "type": "integer",
            "title": "Timeout Seconds",
            "default": 30
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.2
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "default": 1024
          },
          "prompt_cache_enabled": {
            "type": "boolean",
            "title": "Prompt Cache Enabled",
            "default": true
          },
          "prompt_cache_ttl": {
            "type": "string",
            "enum": [
              "ephemeral",
              "one_hour"
            ],
            "title": "Prompt Cache Ttl",
            "default": "ephemeral"
          },
          "provider_options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provider Options"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "base_url",
          "api_key",
          "model_name"
        ],
        "title": "ModelConnectionTestRequest"
      },
      "ModelConnectionTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ModelConnectionTestResponse"
      },
      "ModelDiscoveryRequest": {
        "properties": {
          "config_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Id"
          },
          "provider": {
            "type": "string",
            "const": "anthropic",
            "title": "Provider",
            "default": "anthropic"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "timeout_seconds": {
            "type": "integer",
            "title": "Timeout Seconds",
            "default": 30
          },
          "provider_options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provider Options"
          }
        },
        "type": "object",
        "required": [
          "base_url"
        ],
        "title": "ModelDiscoveryRequest"
      },
      "ModelDiscoveryResponse": {
        "properties": {
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelDiscoveryResponse"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}